Follower list logic changed
The bot now updates it's followers list every 4 hours and 20 seconds instead of every time the bot writes a new tweet.
This way, if a user is mentioned, it is only possible for it to be mentioned again 5 hours later, in order to prevent excessive spamming.
Time between tweets has been changed to 1 hour. This may or be not be changed back to 30 minutes in a later commit.
The extra 20 seconds work to prevent conflict between a new tweet generation and the follower array update.